Added additionalFields() and tableAlias() to merge reporting.
This makes sure that if you have additionalTables(), and they have
fields with the same name as tableName(), then the fields from
additionalTables() will not overwrite fields from tableName().

Updated all contacts mergers who needed it so that they use
additionalFields() to specify what to put in the SELECT clause.