Helper branch, for testing the changes in feature/CONTACTS-276-de-dupe-performance.
It stuffs around with error reporting, and puts some wrappers around the invocation of
"reportTitle()" method in ContactsMergeReport, so that if we broke any of them, they
will throw an error due to a php error (NOTICE: Undefined index ...).

Didn't know where else to put it, and was sick of stashing and unstashing.